requirements-gpu.txt
tensorflow-gpu==2.3.0rc0
opencv-python==4.1.1.26
tensorflow==2.2.3
lxml
tqdm
absl-py
matplotlib
easydict
pillow
keras==2.4.3